requirements-spatialflow.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scanpy>=1.9.3
anndata>=0.9.1
squidpy>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# Spatial analysis
ripser>=0.6.1
gudhi>=3.8.0
mgwr>=2.1.0

# Bayesian modelling
pymc3>=3.11.5
gpflow>=2.5.0
tensorflow>=2.10.0

# Visualization
plotly>=5.14.0
networkx>=3.0.0
bokeh>=3.1.0

# CLI and configuration
click>=8.1.3
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.2.0
statsmodels>=0.13.5
numba>=0.56.0
python-igraph>=0.10.0
leidenalg>=0.9.0

# Optional dependencies for advanced features
# pystan>=3.5.0
# tradeseq @ git+https://github.com/statOmics/tradeSeq.git
# cell2location>=0.1.0